Added Railcraft-derived code.
All files are from the Railcraft mod, used in accordance to it's license.  As these files comprise of less than 25% of the code in the project, it is within the constraints of the license clause "You MAY use snippets of Railcraft code in your own projects, but only if your project consists of less than 25% of Railcraft derived code. You must give credit to the Railcraft Project for the code used, including a link to the Github Project. Put this in your class file headers that contain Railcraft code, in your readme, and on the main download page."